pyproject.toml
[project]
name = "hub-dados-ia-es"
version = "0.1.0"
description = "ExtrAI - Hybrid RAG System for Technical Intelligence"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.53.0",
    "langchain>=1.2.7",
    "langchain-core>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=0.3.0",    # Para Azure e Local LM Studio
    "langchain-aws>=0.2.0",       # Integração moderna com Bedrock
    "langchain-chroma>=0.1.2",    # O novo "Cérebro" (Substitui FAISS)
    "langchain-huggingface>=0.1.0", # Para rodar o BGE-M3 localmente
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",              # Essencial para GPU (CDI) e MPS (M3 Pro)
    "httpx>=0.28.1",
    "pandas>=2.2.2,<3.0.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "python-dotenv>=1.0.1",
]

    # uv sync --extra corporate
[project.optional-dependencies]
corporate = [
"iaaws-lib>=0.2.1",
]

[tool.uv]
native-tls = true
allow-insecure-host = ["nexus.petrobras.com.br"]

# 1. PyPI Oficial (Padrão para coisas públicas = Rápido e Estável)
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

# 2. Nexus Petrobras (Secundário = Apenas para coisas internas)
[[tool.uv.index]]
name = "petrobras-nexus"
url = "https://nexus.petrobras.com.br/repository/pypi-all/simple"
explicit = true

# 3. A Regra de Ouro: Força o pacote interno a vir do Nexus
[tool.uv.sources]
iaaws-lib = { index = "petrobras-nexus" }